% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aston_parsers.R
\name{configure_aston}
\alias{configure_aston}
\title{Configure Aston}
\usage{
configure_aston(return_boolean = FALSE)
}
\arguments{
\item{return_boolean}{Logical. Whether to return a Boolean value indicating
if the chromConverter environment is correctly configured.}
}
\value{
If \code{return_boolean} is \code{TRUE}, returns a Boolean value
indicating whether the chromConverter environment is configured correctly.
Otherwise, there is no return value.
}
\description{
Configures reticulate to use Aston file parsers.
}
\author{
Ethan Bass
}
