% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_lcd.R
\name{configure_olefile}
\alias{configure_olefile}
\title{Configure olefile}
\usage{
configure_olefile(return_boolean = FALSE)
}
\arguments{
\item{return_boolean}{Logical. Whether to return a Boolean value indicating
if the chromConverter environment is correctly configured.}
}
\value{
If \code{return_boolean} is \code{TRUE}, returns a Boolean value
indicating whether the chromConverter environment is configured correctly.
Otherwise, there is no return value.
}
\description{
Configures reticulate to use olefile. Olefile is required to use the 'Shimadzu'
LCD parser.
}
\author{
Ethan Bass
}
