% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow_parser.R
\name{configure_rainbow}
\alias{configure_rainbow}
\title{Configure rainbow}
\usage{
configure_rainbow()
}
\value{
No return value.
}
\description{
Configures reticulate to use rainbow file parsers.
}
\author{
Ethan Bass
}
