% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chroms.R
\name{entab_reader}
\alias{entab_reader}
\title{Entab parsers}
\usage{
entab_reader(
  file,
  format_data = c("wide", "long"),
  format_in = "",
  format_out = c("matrix", "data.frame"),
  read_metadata = TRUE
)
}
\arguments{
\item{file}{path to file}

\item{format_data}{Whether to output data in wide or long format.}

\item{format_in}{Format of input.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{read_metadata}{Whether to read metadata from file.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Entab parsers
}
