% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-roxygen.R
\name{generic_return_3D}
\alias{generic_return_3D}
\title{Generic return (3D)}
\value{
A 3D chromatogram in the format specified by \code{data_format} and
\code{format_out}. If \code{data_format} is \code{wide}, the chromatogram will
be returned with retention times as row names and wavelengths as columns. If
\code{long} format is requested, three columns will be returned: one for the
retention time, one for the wavelength and one for the intensity. The
\code{format_out} argument determines whether the chromatogram is returned as
a \code{matrix}, \code{data.frame}, or \code{data.table}. Metadata are attached to the
chromatogram as \link{attributes} if \code{read_metadata} is \code{TRUE}.
}
\description{
Generic return (3D)
}
\keyword{internal}
