% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chroms.R
\name{import_chroms}
\alias{import_chroms}
\title{import_chroms}
\usage{
import_chroms(
  paths,
  pattern = NULL,
  format.in = c("chemstation.uv", "masshunter.dad"),
  R.format = c("matrix", "data.frame"),
  export = FALSE,
  path.out = NULL,
  format.out = "csv",
  dat = NULL
)
}
\arguments{
\item{paths}{paths to folders containing files}

\item{pattern}{pattern (e.g. a file extension). Defaults to NULL, in which
case file extension will be deduced from \code{format.in}.}

\item{format.in}{Format of files to be imported/converted.}

\item{R.format}{R object format (i.e. data.frame or matrix).}

\item{export}{Logical. If true, will export files as csvs.}

\item{path.out}{Path for export.}

\item{format.out}{Output format. Currently only .csv.}

\item{dat}{Existing list of chromatographs to append results.
(Defaults to NULL).}
}
\value{
A list of chromatograms in matrix or data.frame format.
}
\description{
import_chroms
}
\examples{
\dontrun{
data <- import_chrom(paths)
}
}
\author{
Ethan Bass
}
