% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_dependencies.R
\name{install_python_dependencies}
\alias{install_python_dependencies}
\title{Basic function to install python dependencies.}
\usage{
install_python_dependencies(...)
}
\arguments{
\item{...}{Additional commands to \code{\link[reticulate]{py_install}}}
}
\description{
Basic function to install python dependencies.
}
