% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_thermoraw.R
\name{mzML_UV_parser}
\alias{mzML_UV_parser}
\title{Extract UV data from mzML files}
\usage{
mzML_UV_parser(path)
}
\arguments{
\item{path}{path to file}
}
\value{
A chromatograms in \code{matrix} format.
}
\description{
Extracts UV data from mzML files
}
\author{
Ethan Bass
}
