% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openchrom_parser.R
\name{openchrom_parser}
\alias{openchrom_parser}
\title{Parse files with OpenChrom}
\usage{
openchrom_parser(
  files,
  path_out,
  format_in,
  format_out = c("mzml", "cdf", "animl", "csv")
)
}
\arguments{
\item{files}{files to parse}

\item{path_out}{directory to export converted files.}

\item{format_in}{Either \code{msd} for mass spectrometry data, \code{csd} for flame ionization data, or \code{wsd} for DAD/UV data.}

\item{format_out}{Either \code{mzml}, \code{cdf}, \code{animl}, or \code{csv}.}
}
\value{
No return value.
}
\description{
To use this function \href{https://lablicate.com/platform/openchrom}{OpenChrom} must be manually installed.
}
\author{
Ethan Bass
}
