% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_fid.R
\name{read_chemstation_fid}
\alias{read_chemstation_fid}
\title{Parser for reading Agilent FID (.ch) files into R}
\usage{
read_chemstation_fid(
  path,
  read_metadata = TRUE,
  format_out = c("matrix", "data.frame")
)
}
\arguments{
\item{path}{Path to \code{.ch} file}

\item{read_metadata}{Logical. Whether to attach metadata.}

\item{format_out}{Matrix or data.frame}
}
\description{
Parser for reading Agilent FID (.ch) files into R
}
\note{
This function benefited greatly from the schematic shared by the developers
of the rainbow project (https://rainbow-api.readthedocs.io/en/latest/agilent/ch_fid.html).
}
\author{
Ethan Bass
}
