% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_ms.R
\name{read_chemstation_ms}
\alias{read_chemstation_ms}
\title{Read 'Agilent ChemStation' MS files}
\usage{
read_chemstation_ms(
  path,
  what = c("MS1", "BPC", "TIC"),
  format_out = c("matrix", "data.frame", "data.table"),
  data_format = "long",
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  collapse = TRUE
)
}
\arguments{
\item{path}{Path to 'Agilent' \code{.ms} file.}

\item{what}{Which streams to return: \code{MS1}, \code{BPC} and/or \code{TIC}. Defaults to
all three.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{data_format}{Whether to return the \code{BPC} and \code{TIC} in \code{long} (default)
or \code{wide} format. Mass spectra are always returned in \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
(standardized field names) or \code{raw} (vendor field names, unmapped).}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element. Defaults to \code{TRUE}.}
}
\value{
A list of the streams in \code{what}, in the format specified by
\code{format_out}. In wide format, the \code{BPC} and \code{TIC} have retention times as
rows and a single intensity column; in long format, a retention time column
and an intensity column. MS data will always be returned in long format. The
\code{format_out} argument determines whether the chromatogram is returned as a
\code{matrix}, \code{data.frame}, or \code{data.table}. Metadata are attached as \link{attributes} if
\code{read_metadata} is \code{TRUE}. With \code{collapse = TRUE}, a list of one stream is
replaced by that stream.
}
\description{
Reads 'Agilent ChemStation MSD Spectral Files' beginning with
\code{x01/x32/x00/x00}.
}
\note{
Many thanks to Evan Shi and Eugene Kwan for providing helpful
information on the structure of these files in the \href{https://rainbow-api.readthedocs.io/en/latest/agilent/ms.html}{rainbow documentation}.
}
\examples{
\dontrun{
read_chemstation_ms(path)
}
}
\seealso{
Other 'Agilent' parsers:
\code{\link[=read_agilent_d]{read_agilent_d()}},
\code{\link[=read_agilent_dx]{read_agilent_dx()}},
\code{\link[=read_agilent_rslt]{read_agilent_rslt()}},
\code{\link[=read_chemstation_ch]{read_chemstation_ch()}},
\code{\link[=read_chemstation_csv]{read_chemstation_csv()}},
\code{\link[=read_chemstation_logs]{read_chemstation_logs()}},
\code{\link[=read_chemstation_reports]{read_chemstation_reports()}},
\code{\link[=read_chemstation_uv]{read_chemstation_uv()}}
}
\author{
Ethan Bass
}
\concept{'Agilent' parsers}
