% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fid.R
\name{read_fid}
\alias{read_fid}
\title{Read Shimadzu txt files from GC-FID into R}
\usage{
read_fid(
  paths,
  suffix = "txt",
  dat = NULL,
  R.format = c("data.frame", "tibble")
)
}
\arguments{
\item{paths}{path to files (or vector of paths)}

\item{suffix}{string specifying file suffix (e.g. 'txt')}

\item{dat}{files Optional list of chromatograms. If provided, newly imported
chromatograms will be appended to the existing list.}

\item{R.format}{R object format (i.e. data.frame or tibble).}
}
\value{
A list of chromatograms in data.frame or tibble format, according to
the value of 'R.format'.
}
\description{
Read Shimadzu txt files from GC-FID into R
}
\examples{
path <- system.file("extdata", ".", package = "chromConverter")
read_fid(path)
}
\author{
Ethan Bass
}
