% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chroms.R
\name{read_shimadzu_fid}
\alias{read_shimadzu_fid}
\title{Shimadzu ascii FID reader}
\usage{
read_shimadzu_fid(
  file,
  read_metadata = TRUE,
  format_out = c("matrix", "data.frame")
)
}
\arguments{
\item{file}{path to file}

\item{read_metadata}{Whether to read metadata from file.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}
}
\value{
A matrix object (retention time x trace).
}
\description{
Shimadzu ascii FID reader
}
\author{
Ethan Bass
}
