% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_lcd.R
\name{read_sz_tic}
\alias{read_sz_tic}
\title{A parser to read total ion chromatogram data streams from 'Shimadzu'
\code{.lcd} files. LCD files are encoded as 'Microsoft' OLE documents. The
parser relies on the \href{https://pypi.org/project/olefile/}{olefile} package in
Python to unpack the files. The TIC data is encoded in a stream called
\code{Centroid:3D Raw Data}.
The PDA data stream contains a segment for each retention time, beginning
with a 8-byte header. After the header, the file consists of a series of
4-byte little-endian integers in blocks of 3 (16-bytes per block), followed by
a 4-byte spacer (\code{00000000}) The first integer is the retention time
(scaled by 1000), the second integer is the scan number, and the third integer
is the intensity.}
\usage{
read_sz_tic(
  path,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  read_metadata = TRUE
)
}
\arguments{
\item{path}{Path to LCD file.}

\item{format_out}{Matrix or data.frame.}

\item{data_format}{Either \code{wide} (default) or \code{long}.}

\item{read_metadata}{Logical. Whether to attach metadata.}
}
\value{
A 2D chromatogram from the SumTIC stream in \code{matrix} or
\code{data.frame} format, according to the value of \code{format_out}.
The chromatograms will be returned in \code{wide} or \code{long} format
according to the value of \code{data_format}.
}
\description{
A parser to read total ion chromatogram data streams from 'Shimadzu'
\code{.lcd} files. LCD files are encoded as 'Microsoft' OLE documents. The
parser relies on the \href{https://pypi.org/project/olefile/}{olefile} package in
Python to unpack the files. The TIC data is encoded in a stream called
\code{Centroid:3D Raw Data}.
The PDA data stream contains a segment for each retention time, beginning
with a 8-byte header. After the header, the file consists of a series of
4-byte little-endian integers in blocks of 3 (16-bytes per block), followed by
a 4-byte spacer (\code{00000000}) The first integer is the retention time
(scaled by 1000), the second integer is the scan number, and the third integer
is the intensity.
}
\note{
This parser is experimental and may still need some work. It is not
yet able to interpret much metadata from the files.
}
\author{
Ethan Bass
}
