% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varian_peaklist.R
\name{read_varian_peaklist}
\alias{read_varian_peaklist}
\title{Read 'Varian' peak list}
\usage{
read_varian_peaklist(path)
}
\arguments{
\item{path}{Path to 'Varian' peak list file.}
}
\value{
A \code{data.frame} containing the information from the specified report.
}
\description{
Read peak list(s) from 'Varian MS Workstation'.
}
\examples{
\dontrun{
read_varian_peaklist(path)
}
}
\seealso{
Other 'Varian' parsers:
\code{\link[=read_varian_sms]{read_varian_sms()}}
}
\author{
Ethan Bass
}
\concept{'Varian' parsers}
