% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_waters_raw.R
\name{read_waters_raw}
\alias{read_waters_raw}
\title{Read 'Waters' RAW}
\usage{
read_waters_raw(
  path,
  format_out = c("matrix", "data.frame", "data.table"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw")
)
}
\arguments{
\item{path}{Path to Waters \code{.raw} file.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{data_format}{Whether to return data in \code{wide} (default) or \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
(standardized field names) or \code{raw} (vendor field names, unmapped).}
}
\value{
A chromatogram in the format specified by the \code{format_out} and
\code{data_format} arguments.
}
\description{
Reads 'Waters MassLynx' (\code{.raw}) files into R.
}
\note{
For now this parser only reads 1D chromatograms (not mass spectra or
DAD data) and does not support parsing of metadata from 'Waters' RAW files.
}
\examples{
\dontrun{
read_waters_raw("path/to/file.raw")
}
}
\seealso{
Other 'Waters' parsers:
\code{\link[=read_waters_arw]{read_waters_arw()}}
}
\author{
Ethan Bass
}
\concept{'Waters' parsers}
