% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_andi_ms.R
\name{write_andi_ms}
\alias{write_andi_ms}
\title{Write ANDI MS CDF file from chromatogram}
\usage{
write_andi_ms(
  x,
  path_out,
  sample_name = NULL,
  force = FALSE,
  ms_params = list(ionization_mode = "Electron Impact", ionization_polarity =
    "Positive Polarity", detector_type = "Electron Multiplier")
)
}
\arguments{
\item{x}{A list of chromatograms containing an element whose name includes
\code{MS} (e.g. \code{MS1}), or a single chromatogram of MS1 scans, in \code{wide} or
\code{long} format.}

\item{path_out}{The path to write the file.}

\item{sample_name}{The name of the file. If a name is not provided, the name
will be derived from the \code{sample_name} attribute.}

\item{force}{Whether to overwrite existing files at the specified path.
Defaults to \code{FALSE}.}

\item{ms_params}{A list of instrument settings recorded in the file, since
they cannot be derived from the data: \code{ionization_mode},
\code{ionization_polarity} and \code{detector_type}.}
}
\value{
Invisibly returns the path to the written CDF file.
}
\description{
Exports mass spectrometry data in ANDI (Analytical Data Interchange) MS
format (ASTM E1947-98). Retention times are converted to seconds, as the
specification requires. A total ion chromatogram is derived from the MS1
scans if the supplied object does not already carry one.
}
\section{Side effects}{

Exports mass spectrometry data in ANDI MS format (netCDF) in the directory
specified by \code{path_out}. The file will be named according to the value
of \code{sample_name}. If no \code{sample_name} is provided, the \code{sample_name}
attribute will be used if it exists.
}

\examples{
\dontrun{
write_andi_ms(chrom, path_out = "path/to/directory")
}
}
\seealso{
Other write functions:
\code{\link[=write_andi_chrom]{write_andi_chrom()}},
\code{\link[=write_chroms]{write_chroms()}},
\code{\link[=write_mzml]{write_mzml()}}
}
\author{
Ethan Bass
}
\concept{write functions}
