% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_chroms.R
\name{write_cdf}
\alias{write_cdf}
\title{Write CDF file from chromatogram}
\usage{
write_cdf(x, path_out, sample_name, lambda = NULL, force = FALSE)
}
\arguments{
\item{x}{A chromatogram in (wide) format.}

\item{path_out}{The path to write the file.}

\item{sample_name}{The name of the file.}

\item{lambda}{The wavelength to export (for 2-dimensional chromatograms).
Must be a string matching one the columns in \code{x} or the index of the
column to export.}

\item{force}{Whether to overwrite existing files at the specified path.
Defaults to \code{FALSE}.}
}
\value{
No return value. The function is called for its side effects.
}
\description{
Exports a chromatogram in ANDI (Analytical Data Interchange) chromatography
format (ASTM E1947-98). This format can only accomodate unidimensional data.
For two-dimensional chromatograms, the column to export can be specified
using the \code{lambda} argument. Othewise, a warning will be generated and
the first column of the chromatogram will be exported.
}
\section{Side effects}{

Exports a chromatogram in ANDI chromatography format (netCDF) in the directory
specified by \code{path_out}. The file will be named according to the value
of \code{sample_name}. If no \code{sample_name} is provided, the
\code{sample_name} attribute will be used if it exists.
}

\author{
Ethan Bass
}
