% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_openchrom.R
\name{write_openchrom_batchfile}
\alias{write_openchrom_batchfile}
\title{Writes OpenChrom XML batch file
This function is called internally by \code{call_openchrom}.}
\usage{
write_openchrom_batchfile(
  files,
  path_out,
  format_in = c("msd", "csd", "wsd"),
  export_format = c("csv", "cdf", "mzml", "animl")
)
}
\arguments{
\item{files}{Paths to files for conversion}

\item{path_out}{directory to export converted files.}

\item{format_in}{Either \code{msd} for mass spectrometry data, \code{csd} for
flame ionization data, or \code{wsd} for DAD/UV data.}

\item{export_format}{Either \code{csv}, \code{cdf}, \code{mzml}, or \code{animl}.}
}
\value{
Returns path to newly created xml batch file.
}
\description{
Writes OpenChrom XML batch file
This function is called internally by \code{call_openchrom}.
}
\author{
Ethan Bass
}
