% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_metadata}
\alias{attach_metadata}
\title{Attaches metadata to chromatogram}
\usage{
attach_metadata(
  x,
  meta,
  format_in,
  format_out,
  data_format,
  parser = NULL,
  source_file
)
}
\arguments{
\item{x}{chromatogram}

\item{meta}{List object containing metadata.}

\item{format_in}{Chromatogram format}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{data_format}{Whether data is in wide or long format.}

\item{parser}{What parser was used to decode the data.}

\item{source_file}{The path to the source file.}
}
\value{
A chromatogram with attached metadata.
}
\description{
Attaches metadata to chromatogram
}
\author{
Ethan Bass
}
