% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_metadata}
\alias{attach_metadata}
\title{Attaches metadata to chromatogram}
\usage{
attach_metadata(x, meta, format_in, format_out, format_data, parser = NULL)
}
\arguments{
\item{x}{chromatogram}

\item{meta}{List object containing metadata.}

\item{format_in}{Chromatogram format}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{format_data}{Whether data is in wide or long format.}

\item{parser}{What parser was used to decode the data}
}
\value{
A chromatogram with attached metadata.
}
\description{
Attaches metadata to chromatogram
}
\author{
Ethan Bass
}
