% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_entab.R
\name{call_entab}
\alias{call_entab}
\title{Call Entab}
\usage{
call_entab(
  path,
  data_format = c("wide", "long"),
  format_out = c("matrix", "data.frame", "data.table"),
  format_in = "",
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw")
)
}
\arguments{
\item{path}{Path to file.}

\item{data_format}{Whether to return data in \code{wide} (default) or \code{long} format.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{format_in}{Format of input.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
(standardized field names) or \code{raw} (vendor field names, unmapped).}
}
\value{
A chromatogram in the format specified by the \code{format_out} and
\code{data_format} arguments.
}
\description{
Converts chromatography date files using \href{https://github.com/bovee/entab}{entab}
parsers.
}
\examples{
\dontrun{
call_entab("path/to/file.uv", format_in = "chemstation_uv")
}
}
\seealso{
Other external parsers:
\code{\link[=call_openchrom]{call_openchrom()}},
\code{\link[=call_rainbow]{call_rainbow()}},
\code{\link[=read_thermoraw]{read_thermoraw()}},
\code{\link[=sp_converter]{sp_converter()}},
\code{\link[=uv_converter]{uv_converter()}}
}
\concept{external parsers}
