% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_entab.R
\name{call_entab}
\alias{call_entab}
\title{Call Entab
Converts files using Entab parsers}
\usage{
call_entab(
  file,
  data_format = c("wide", "long"),
  format_in = "",
  format_out = c("matrix", "data.frame"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw")
)
}
\arguments{
\item{file}{path to file}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.}

\item{format_in}{Format of input.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{read_metadata}{Whether to read metadata from file.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Call Entab
Converts files using Entab parsers
}
