% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{configure_aston}
\alias{configure_aston}
\title{Configure Aston}
\usage{
configure_aston()
}
\value{
No return value.
}
\description{
Configures reticulate to use Aston file parsers.
}
\author{
Ethan Bass
}
