% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_openchrom.R
\name{configure_openchrom}
\alias{configure_openchrom}
\title{Configure 'OpenChrom' parser}
\usage{
configure_openchrom(cli = c("null", "true", "false", "status"), path = NULL)
}
\arguments{
\item{cli}{Defaults to NULL. If "true", R will rewrite openchrom ini file to enable CLI.
If "false", R will disable CLI. If NULL, R will not modify the ini file.}

\item{path}{Path to 'OpenChrom' executable (Optional). The supplied path will
overwrite the current path.}
}
\value{
If \code{cli} is set to \code{"status"}, returns a Boolean value
indicating whether 'OpenChrom' is configured correctly. Otherwise, returns
the path to OpenChrom command-line application.
}
\description{
Configures \href{https://lablicate.com/platform/openchrom}{OpenChrom} to use
command-line interface. Requires OpenChrom version prior to 1.5.0.
}
\examples{
\dontrun{
configure_openchrom(cli = "status")
}
}
\seealso{
\code{\link{call_openchrom}}
}
\author{
Ethan Bass
}
