% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_python_environment}
\alias{configure_python_environment}
\title{Configure python environment}
\usage{
configure_python_environment(
  what = c("venv", "conda"),
  envname = "chromConverter",
  python = reticulate::virtualenv_starter(),
  ...
)
}
\arguments{
\item{what}{What kind of virtual environment to create. A python virtual
environment (\code{"venv"}) or a conda environment (\code{"conda"}).}

\item{envname}{The name of, or path to, a Python virtual environment.}

\item{python}{Argument to \code{reticulate::virtualenv_create}, specifying
the path to a Python interpreter.}

\item{...}{Additional arguments to \link[reticulate:virtualenv-tools]{reticulate::virtualenv_create} or
\link[reticulate:conda-tools]{reticulate::conda_create} according to the value of \code{what}.}
}
\value{
There is no return value.
}
\description{
Configures python virtual environment or conda environment for parsers that
have python dependencies, according to the value of \code{what}. While this
should not be necessary in most cases starting with reticulate \code{v1.41.0},
this function can be used to create a dedicated chromConverter environment.
}
\section{Side effects}{

Creates and configures either  a python virtual environment or conda
environment (according to the value of \code{what}) with all the packages
required for running chromConverter.
}

\author{
Ethan Bass
}
\keyword{internal}
