% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_python_environment}
\alias{configure_python_environment}
\title{Configure python environment}
\usage{
configure_python_environment(parser = "all", return_boolean = FALSE)
}
\arguments{
\item{parser}{Either \code{aston}, \code{rainbow}, or \code{olefile} (for
\code{read_shimadzu_lcd}).}

\item{return_boolean}{Logical. Whether to return a Boolean value indicating
if the chromConverter environment is correctly configured.}
}
\value{
If \code{return_boolean} is \code{TRUE}, returns a Boolean value
indicating whether the chromConverter environment is configured correctly.
Otherwise, there is no return value.
}
\description{
Configures reticulate environment for parsers that have python dependencies.
Deprecated as this should no longer be necessary with reticulate \code{v1.41.0}.
}
\author{
Ethan Bass
}
\keyword{internal}
