% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata}
\usage{
extract_metadata(
  chrom_list,
  what = c("instrument", "detector", "software", "method", "batch", "operator",
    "run_date", "sample_name", "sample_id", "injection_volume", "time_range",
    "time_interval", "detector_range", "data_format", "parser", "format_out"),
  format_out = c("data.frame", "tibble")
)
}
\arguments{
\item{chrom_list}{A list of chromatograms with attached metadata (as returned
by \code{read_chroms} with \code{read_metadata = TRUE}).}

\item{what}{A character vector specifying the metadata elements to extract.}

\item{format_out}{Format of object. Either \code{data.frame} or \code{tibble}.}
}
\value{
A data.frame or tibble (according to the value of \code{format_out}),
with samples as rows and the specified metadata elements as columns.
}
\description{
Extract metadata as a data.frame from a list of chromatograms.
}
