% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata}
\usage{
extract_metadata(
  chrom_list,
  what = c("instrument", "detector", "detector_id", "software", "method", "batch",
    "operator", "run_datetime", "sample_name", "sample_id", "injection_volume",
    "time_range", "time_interval", "time_unit", "detector_range", "detector_y_unit",
    "detector_x_unit", "intensity_multiplier", "scaled", "source_file",
    "source_file_format", "source_sha1", "data_format", "parser", "format_out"),
  format_out = c("data.frame", "data.table", "tibble")
)
}
\arguments{
\item{chrom_list}{A list of chromatograms with attached metadata (as returned
by \code{read_chroms} with \code{read_metadata = TRUE}).}

\item{what}{A character vector specifying the metadata elements to extract.}

\item{format_out}{Format of object. Either \code{data.frame}, \code{tibble},
or \code{data.table}.}
}
\value{
A \code{data.frame}, \code{tibble}, or \code{data.table} (according
to the value of \code{format_out}), with samples as rows and the specified
metadata elements as columns.
}
\description{
Extract metadata as a \code{data.frame} or \code{tibble} from a list of
chromatograms.
}
