% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-roxygen.R
\name{generic_return_2D}
\alias{generic_return_2D}
\title{Generic return (2D)}
\value{
A 2D chromatogram in the format specified by \code{data_format} and
\code{format_out}. If \code{data_format} is \code{wide}, the chromatogram will be returned
with retention times as rows and a single column for the intensity. If \code{long}
format is requested, two columns will be returned: one for the
retention time and one for the intensity. The \code{format_out} argument
determines whether the chromatogram is returned as a \code{matrix}, \code{data.frame},
or \code{data.table}. Metadata can be attached to the chromatogram as \link{attributes}
if \code{read_metadata} is \code{TRUE}.
}
\description{
Generic return (2D)
}
\keyword{internal}
