% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chrom.R
\name{import_chrom}
\alias{import_chrom}
\title{import_chrom}
\usage{
import_chrom(
  paths,
  pattern = ".uv",
  dat = NULL,
  format = c("matrix", "data.frame"),
  export = FALSE
)
}
\arguments{
\item{paths}{paths to folders containing files}

\item{pattern}{pattern (e.g. a file extension)}

\item{dat}{existing list of chromatographs to append results.
(Defaults to NULL).}

\item{format}{data.frame or matrix}

\item{export}{Logical. If true, will export files as csvs.}
}
\description{
import_chrom
}
\examples{
\dontrun{
data <- import_UV(paths)}
}
