% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openchrom_parser.R
\name{openchrom_parser}
\alias{openchrom_parser}
\title{Parse files with OpenChrom}
\usage{
openchrom_parser(
  files,
  path_out,
  format_in,
  export_format = c("mzml", "cdf", "animl", "csv"),
  return_paths = FALSE
)
}
\arguments{
\item{files}{files to parse}

\item{path_out}{directory to export converted files.}

\item{format_in}{Either \code{msd} for mass spectrometry data, \code{csd} for flame ionization data, or \code{wsd} for DAD/UV data.}

\item{export_format}{Either \code{mzml}, \code{cdf}, \code{animl}, or \code{csv}.}

\item{return_paths}{Logical. If TRUE, the function will return a character vector of paths to the newly created files.}
}
\value{
If \code{return_paths} is TRUE, the function will return a vector of paths to the newly created files.
If \code{return_paths} is FALSE and \code{export_format} is \code{csv}, the function will return a list
of chromatograms in \code{data.frame} format. Otherwise, it will not return anything.
}
\description{
To use this function \href{https://lablicate.com/platform/openchrom}{OpenChrom} must be manually installed.
}
\details{
The \code{openchrom_parser} works by creating an \code{xml} batchfile and
feeding it to the OpenChrom commandline interface. OpenChrom batchfiles
consist of \code{InputEntries} (the files you want to convert) and \code{
ProcessEntries} (what you want to do to the files). The parsers are organized
into broad categories by detector-type and output format. The detector-types
are \code{msd} (mass selective detectors), \code{csd} (current selective
detectors, such as FID, ECD, NPD), and \code{wsd} (wavelength selective
detectors, such as  DAD, and UV/VIS). Thus, when calling the OpenChrom parsers,
you must select one of these three options for the input format (\code{format_in}).
}
\section{Side effects}{
 Chromatograms will be exported in the format specified
by \code{export_format} in the folder specified by \code{path_out}.
}

\author{
Ethan Bass
}
