% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agilent_dx.R
\name{read_agilent_dx}
\alias{read_agilent_dx}
\title{Read 'Agilent' DX files}
\usage{
read_agilent_dx(
  path,
  what = c("chroms", "dad"),
  path_out = NULL,
  format_out = c("matrix", "data.frame", "data.table"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  collapse = TRUE
)
}
\arguments{
\item{path}{Path to Agilent \code{.dx} file.}

\item{what}{Whether to extract chromatograms (\code{chroms}), DAD data
(\code{dad}) and/or auxiliary instrumental data (\code{instrument}) (e.g.,
temperature, pressure, solvent composition, etc.). Accepts multiple arguments.}

\item{path_out}{A directory to export unzipped files. If a path is not
specified, the files will be written to a temp directory on the disk. The
function will overwrite existing folders in the specified directory
that share the basename of the file specified by \code{path}.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{data_format}{Whether to return data in \code{wide} (default) or \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
(standardized field names) or \code{raw} (vendor field names, unmapped).}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element. Defaults to \code{TRUE}.}
}
\value{
A chromatogram in the format specified by \code{format_out} (retention
time x wavelength).
}
\description{
Reads 'Agilent' \code{.dx} files.
}
\details{
This function unzips 'Agilent'  \code{.dx} into a temporary directory using
\link{unzip} and calls the appropriate parser on the unzipped file.
}
\examples{
\dontrun{
read_agilent_dx(path)
}
}
\seealso{
Other 'Agilent' parsers:
\code{\link[=read_agilent_d]{read_agilent_d()}},
\code{\link[=read_agilent_rslt]{read_agilent_rslt()}},
\code{\link[=read_chemstation_ch]{read_chemstation_ch()}},
\code{\link[=read_chemstation_csv]{read_chemstation_csv()}},
\code{\link[=read_chemstation_ms]{read_chemstation_ms()}},
\code{\link[=read_chemstation_reports]{read_chemstation_reports()}},
\code{\link[=read_chemstation_uv]{read_chemstation_uv()}}
}
\author{
Ethan Bass
}
\concept{'Agilent' parsers}
