% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agilent_dx.R
\name{read_agilent_dx}
\alias{read_agilent_dx}
\title{Read 'Agilent' DX files}
\usage{
read_agilent_dx(
  path,
  what = c("chroms", "dad"),
  path_out = NULL,
  format_out = c("matrix", "data.frame", "data.table"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  collapse = TRUE
)
}
\arguments{
\item{path}{Path to \code{.dx} file.}

\item{what}{Whether to extract chromatograms (\code{chroms}), DAD data
(\code{dad}) and/or auxiliary instrumental data \code{instrument} (e.g.,
temperature, pressure, solvent composition, etc.). Accepts multiple arguments.}

\item{path_out}{A directory to export unzipped files. If a path is not
specified, the files will be written to a temp directory on the disk. The
function will overwrite existing folders in the specified directory
that share the basename of the file specified by \code{path}.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame},
or \code{data.table}.}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Reads 'Agilent' \code{.dx} files.
}
\details{
This function unzips 'Agilent'  \code{.dx} into a temporary directory using
\code{\link{unzip}} and calls \code{\link{read_chemstation_ch}}.
}
\seealso{
Other 'Agilent' parsers: 
\code{\link{read_agilent_d}()},
\code{\link{read_chemstation_ch}()},
\code{\link{read_chemstation_csv}()},
\code{\link{read_chemstation_ms}()},
\code{\link{read_chemstation_reports}()},
\code{\link{read_chemstation_uv}()}
}
\author{
Ethan Bass
}
\concept{'Agilent' parsers}
