% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_ch.R
\name{read_agilent_dx}
\alias{read_agilent_dx}
\title{Parser for reading Agilent ('.dx') files into R}
\usage{
read_agilent_dx(
  path,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  read_metadata = TRUE
)
}
\arguments{
\item{path}{Path to \code{.dx} file}

\item{format_out}{Matrix or data.frame.}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Parser for reading Agilent ('.dx') files into R
}
\author{
Ethan Bass
}
