% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdf.R
\name{read_cdf}
\alias{read_cdf}
\title{Read CDF file}
\usage{
read_cdf(
  file,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  what = "chromatogram",
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  collapse = TRUE
)
}
\arguments{
\item{file}{Path to file.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.
For 2D files, "long" format returns the retention time as the first column of
the data.frame or matrix while "wide" format returns the retention time as the
rownames of the object.}

\item{what}{Whether to extract \code{chromatogram} and/or \code{peak_table}.}

\item{read_metadata}{Whether to read metadata from file.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element.}
}
\value{
A chromatogram in the format specified by the \code{format_out} and
\code{data_format} arguments (retention time x wavelength).
}
\description{
Read CDF file
}
\author{
Ethan Bass
}
