% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_ch.R
\name{read_chemstation_ch}
\alias{read_chemstation_ch}
\title{Read 'Agilent ChemStation' CH files}
\usage{
read_chemstation_ch(
  path,
  format_out = c("matrix", "data.frame", "data.table"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  scale = TRUE,
  source_file = NULL
)
}
\arguments{
\item{path}{Path to 'Agilent' \code{.ch} file.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{data_format}{Whether to return data in \code{wide} (default) or \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
(standardized field names) or \code{raw} (vendor field names, unmapped).}

\item{scale}{Whether to multiply the data by the file's scaling factor and
add its intercept. Defaults to \code{TRUE}. 'MassHunter' seems to ignore the
scaling factor in at least some types of 'ChemStation' files.}

\item{source_file}{Source file from which chromatogram data was originally
derived.}
}
\value{
A 2D chromatogram in the format specified by \code{data_format} and
\code{format_out}. If \code{data_format} is \code{wide}, the chromatogram will be returned
with retention times as row names and a single column for the intensity. If
\code{long} format is requested, two columns will be returned: one for the
retention time and one for the intensity. The \code{format_out} argument
determines whether the chromatogram is returned as a \code{matrix}, \code{data.frame},
or \code{data.table}. Metadata are attached to the chromatogram as \link{attributes}
if \code{read_metadata} is \code{TRUE}.
}
\description{
Reads 'Agilent ChemStation' \code{.ch} files.
}
\details{
'Agilent' \code{.ch} files come in several formats. This parser detects the
version from the file and reads versions \code{8}, \code{30}, \code{130}, \code{81}, \code{179} and
\code{181} from 'Agilent ChemStation' and 'Agilent OpenLab'; any other version is
an error. Versions \code{30} and \code{130} are generally produced by ultraviolet
detectors, and \code{81}, \code{179} and \code{181} by flame ionization detectors (FID).
}
\note{
This function was adapted from the \href{https://github.com/chemplexity/chromatography}{Chromatography Toolbox}
(© James Dillon 2014).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
read_chemstation_ch("tests/testthat/testdata/chemstation_130.ch")
\dontshow{\}) # examplesIf}
}
\seealso{
Other 'Agilent' parsers:
\code{\link[=read_agilent_d]{read_agilent_d()}},
\code{\link[=read_agilent_dx]{read_agilent_dx()}},
\code{\link[=read_agilent_rslt]{read_agilent_rslt()}},
\code{\link[=read_chemstation_csv]{read_chemstation_csv()}},
\code{\link[=read_chemstation_logs]{read_chemstation_logs()}},
\code{\link[=read_chemstation_ms]{read_chemstation_ms()}},
\code{\link[=read_chemstation_reports]{read_chemstation_reports()}},
\code{\link[=read_chemstation_uv]{read_chemstation_uv()}}
}
\author{
Ethan Bass
}
\concept{'Agilent' parsers}
