% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_ch.R
\name{read_chemstation_ch}
\alias{read_chemstation_ch}
\title{Parser for reading Agilent ('.ch') files into R}
\usage{
read_chemstation_ch(
  path,
  read_metadata = TRUE,
  format_out = c("matrix", "data.frame")
)
}
\arguments{
\item{path}{Path to \code{.ch} file}

\item{read_metadata}{Logical. Whether to attach metadata.}

\item{format_out}{Matrix or data.frame}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Parser for reading Agilent ('.ch') files into R
}
\note{
This function was adapted from the \href{https://github.com/chemplexity/chromatography}{Chromatography Toolbox}
((c) James Dillon 2014).
}
\author{
Ethan Bass
}
