% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_csv.R
\name{read_chemstation_csv}
\alias{read_chemstation_csv}
\title{Read Chemstation CSV}
\usage{
read_chemstation_csv(
  path,
  format_out = c("matrix", "data.frame", "data.table")
)
}
\arguments{
\item{path}{Path to file}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame},
or \code{data.table}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Reads 'Agilent Chemstation' CSV files into R.
}
\details{
'Agilent Chemstation' CSV files are encoded in UTF-16.
}
\author{
Ethan Bass
}
