% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_chemstation_csv}
\alias{read_chemstation_csv}
\title{Chemstation CSV reader}
\usage{
read_chemstation_csv(file, format_out = c("matrix", "data.frame"))
}
\arguments{
\item{file}{path to file}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Chemstation CSV reader
}
\author{
Ethan Bass
}
