% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_report.R
\name{read_chemstation_reports}
\alias{read_chemstation_reports}
\title{Read Agilent Chemstation Reports}
\usage{
read_chemstation_reports(
  files,
  data_format = c("chromatographr", "chemstation"),
  metadata_format = c("list", "data.frame")
)
}
\arguments{
\item{files}{Paths to Chemstation report files.}

\item{data_format}{Format to output data. Either \code{chromatographr} or
\code{chemstation}.}

\item{metadata_format}{Format to output metadata. Either \code{list} or
\code{data.frame}.}
}
\description{
Read Agilent Chemstation Reports
}
\author{
Ethan Bass
}
