% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_uv.R
\name{read_chemstation_uv}
\alias{read_chemstation_uv}
\title{Parser for reading Agilent UV (.uv) files into R}
\usage{
read_chemstation_uv(
  path,
  read_metadata = TRUE,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long")
)
}
\arguments{
\item{path}{Path to \code{.uv} file}

\item{read_metadata}{Logical. Whether to attach metadata.}

\item{format_out}{Matrix or data.frame}

\item{data_format}{Either \code{wide} (default) or \code{long}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Parser for reading Agilent UV (.uv) files into R
}
\note{
This function was adapted from the parser in the rainbow project
licensed under GPL 3 by Evan Shi
(https://rainbow-api.readthedocs.io/en/latest/agilent/uv.html).
}
\author{
Ethan Bass
}
