% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemstation_uv.R
\name{read_chemstation_uv}
\alias{read_chemstation_uv}
\title{Read 'ChemStation' DAD files}
\usage{
read_chemstation_uv(
  path,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw")
)
}
\arguments{
\item{path}{Path to \code{.uv} file.}

\item{format_out}{Matrix or data.frame.}

\item{data_format}{Either \code{wide} (default) or \code{long}.}

\item{read_metadata}{Logical. Whether to attach metadata.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}
}
\value{
A 3D chromatogram in the format specified by \code{data_format} and
\code{format_out}. If \code{data_format} is \code{wide}, the chromatogram will
be returned with retention times as rows and wavelengths as columns. If
\code{long} format is requested, three columns will be returned: one for the
retention time, one for the wavelength and one for the intensity. The
\code{format_out} argument determines whether the chromatogram is returned as
a \code{matrix} or \code{data.frame}. Metadata can be attached to the
chromatogram as \code{\link{attributes}} if \code{read_metadata} is {TRUE}.
}
\description{
Agilent \code{.uv} files come in several different varieties. This parser can
automatically detect and read several versions of these files from
'Agilent ChemStation' and 'Agilent OpenLab', including versions \code{31} and
\code{131}.
}
\note{
This function was adapted from the parser in the rainbow project
licensed under GPL 3 by Evan Shi
\url{https://rainbow-api.readthedocs.io/en/latest/agilent/uv.html}.
}
\author{
Ethan Bass
}
