% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chromatotec.R
\name{read_chromatotec}
\alias{read_chromatotec}
\title{Read 'Chromatotec' file}
\usage{
read_chromatotec(
  path,
  what = c("chrom", "peak_table"),
  format_out = c("matrix", "data.frame", "data.table"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  collapse = TRUE
)
}
\arguments{
\item{path}{The path to 'Chromatotec' \code{.Chrom} file.}

\item{what}{Whether to extract chromatograms (\code{chrom}) and/or \code{peak_table}
data. Accepts multiple arguments.}

\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{data_format}{Whether to return data in \code{wide} (default) or \code{long} format.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
(standardized field names) or \code{raw} (vendor field names, unmapped).}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element. Defaults to \code{TRUE}.}
}
\value{
A chromatogram and/or peak table from the specified \code{path}, according
to the value of \code{what}. Chromatograms are returned in the format specified by
\code{format_out}.
}
\description{
Reads 'Chromatotec' \code{.Chrom} files.
}
\examples{
\dontrun{
read_chromatotec(path)
}
}
\author{
Ethan Bass
}
