% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_chromeleon}
\alias{read_chromeleon}
\title{Chromeleon ASCII reader}
\usage{
read_chromeleon(
  file,
  format_out = c("matrix", "data.frame"),
  read_metadata = TRUE
)
}
\arguments{
\item{file}{path to file}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{read_metadata}{Whether to read metadata from file.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Chromeleon ASCII reader
}
\author{
Ethan Bass
}
