% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chroms.R
\name{read_chroms}
\alias{read_chroms}
\title{Read Chromatograms}
\usage{
read_chroms(
  paths,
  find_files = TRUE,
  format.in = c("chemstation_uv", "masshunter_dad", "shimadzu_fid", "chromeleon_uv",
               "thermoraw", "mzml"),
  pattern = NULL,
  parser = c("aston", "entab"),
  R.format = c("matrix", "data.frame"),
  export = FALSE,
  path.out = NULL,
  format.out = "csv",
  read_metadata = TRUE,
  dat = NULL
)
}
\arguments{
\item{paths}{paths to files or folders containing files}

\item{find_files}{Logical. Set to \code{TRUE} (default) if you are providing
the function with a folder or vector of folders containing the files.
Otherwise, set to\code{FALSE}.}

\item{format.in}{Format of files to be imported/converted. The current options
are: \code{chemstation_uv}, \code{masshunter_dad}, \code{shimadzu_fid}, or
\code{chromeleon_uv}.}

\item{pattern}{pattern (e.g. a file extension). Defaults to NULL, in which
case file extension will be deduced from \code{format.in}.}

\item{parser}{What parser to use. Current option are \code{aston} or \code{
entab}. Entab must be manually installed from github.}

\item{R.format}{R object format (i.e. data.frame or matrix).}

\item{export}{Logical. If TRUE, will export files as csvs.}

\item{path.out}{Path for exporting files. If path not specified, files will
export to current working directory.}

\item{format.out}{Output format. Currently only \code{.csv}.}

\item{read_metadata}{Logical, whether to attach metadata (if it's available).
Defaults to TRUE.}

\item{dat}{Existing list of chromatograms to append results.
(Defaults to NULL).}
}
\value{
A list of chromatograms in matrix or data.frame format, according to
the value of 'R.format'.
}
\description{
Reads chromatograms from specified folders or vector of paths using file parsers
from \href{https://github.com/bovee/aston}{Aston}, \href{https://github.com/bovee/entab}{Entab},
and \href{https://github.com/compomics/ThermoRawFileParser}{ThermoRawFileParser}.
}
\details{
Currently recognizes Agilent ChemStation '.uv', MassHunter '.dad' files, and
ThermoRaw files.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- "tests/testthat/testdata/dad1.uv"
chr <- read_chroms(path, find_files = FALSE, format.in = "chemstation_uv")
\dontshow{\}) # examplesIf}
}
\author{
Ethan Bass
}
