% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chroms.R
\name{read_chroms}
\alias{read_chroms}
\title{read_chroms}
\usage{
read_chroms(
  paths,
  format.in = c("chemstation.uv", "masshunter.dad"),
  pattern = NULL,
  parser = c("Aston"),
  R.format = c("matrix", "data.frame"),
  export = FALSE,
  path.out = NULL,
  format.out = "csv",
  dat = NULL
)
}
\arguments{
\item{paths}{paths to folders containing files}

\item{format.in}{Format of files to be imported/converted.}

\item{pattern}{pattern (e.g. a file extension). Defaults to NULL, in which
case file extension will be deduced from \code{format.in}.}

\item{parser}{= What parser to use. Currently, the only option is \code{Aston}.}

\item{R.format}{R object format (i.e. data.frame or matrix).}

\item{export}{Logical. If true, will export files as csvs.}

\item{path.out}{Path for exporting files. If path not specified, files will
export to current working directory.}

\item{format.out}{Output format. Currently only \code{.csv}.}

\item{dat}{Existing list of chromatographs to append results.
(Defaults to NULL).}
}
\value{
A list of chromatograms in matrix or data.frame format.
}
\description{
read_chroms
}
\examples{
\dontrun{
data <- read_chroms(paths, format.in = "chemstation.uv")
}
}
\author{
Ethan Bass
}
