% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fid.R
\name{read_fid}
\alias{read_fid}
\title{Read Shimadzu txt files from GC-FID into R}
\usage{
read_fid(
  paths,
  suffix = "txt",
  dat = NULL,
  format.out = c("matrix", "data.frame", "tibble")
)
}
\arguments{
\item{paths}{path to files (or vector of paths)}

\item{suffix}{string specifying file suffix (e.g. 'txt')}

\item{dat}{files optional list of chromatograms. If provided, newly imported
chromatograms will be appended to the existing list.}

\item{format.out}{R object format (i.e. matrix, data.frame, or tibble).}
}
\value{
A list of chromatograms in \code{matrix}, \code{data.frame}, or
\code{tibble} format, according to the value of 'format.out'.
}
\description{
Parser for Shimadzu .txt files from GC-FID
}
\examples{
path <- system.file("extdata", ".", package = "chromConverter")
read_fid(path)
}
\author{
Ethan Bass
}
