% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_thermoraw.R
\name{read_mzml}
\alias{read_mzml}
\title{Extract UV data from mzML files}
\usage{
read_mzml(path, format_out = c("matrix", "data.frame"))
}
\arguments{
\item{path}{path to file}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}
}
\value{
A chromatograms in \code{matrix} format.
}
\description{
Extracts UV data from mzML files
}
\author{
Ethan Bass
}
