% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_ascii.R
\name{read_shimadzu}
\alias{read_shimadzu}
\title{Shimadzu ascii reader}
\usage{
read_shimadzu(
  file,
  format_in,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  peaktable_format = c("chromatographr", "original"),
  what = "chromatogram",
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw")
)
}
\arguments{
\item{file}{path to file}

\item{format_in}{Format of files. \code{fid} or \code{dad}.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.}

\item{peaktable_format}{Whether to return peak tables in \code{chromatographr} or
\code{original} format.}

\item{what}{Whether to extract \code{chromatogram} and/or \code{peak_table}.
Accepts multiple arguments.}

\item{read_metadata}{Whether to read metadata from file.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter} or
\code{raw}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Reads 'Shimadzu' ascii files into R. These files are exported from
'Lab Solutions' by right clicking on samples in the sample list and
selecting
}
\author{
Ethan Bass
}
