% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_ascii.R
\name{read_shimadzu}
\alias{read_shimadzu}
\title{Shimadzu ascii reader}
\usage{
read_shimadzu(
  path,
  what = "chromatogram",
  format_in = NULL,
  include = c("fid", "lc", "dad", "uv", "tic"),
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  peaktable_format = c("chromatographr", "original"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  ms_format = c("data.frame", "list"),
  collapse = TRUE,
  scale = TRUE
)
}
\arguments{
\item{path}{Path to file.}

\item{what}{Whether to extract \code{chromatogram}, \code{peak_table}, and/or
\code{ms_spectra}. Accepts multiple arguments.}

\item{format_in}{This argument is deprecated and is no longer required.}

\item{include}{Which chromatograms to include. Options are \code{fid},
\code{dad}, \code{uv}, \code{tic}, and \code{status}.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.}

\item{peaktable_format}{Whether to return peak tables in \code{chromatographr}
or \code{original} format.}

\item{read_metadata}{Whether to read metadata from file.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}

\item{ms_format}{Whether to return mass spectral data as a (long)
\code{data.frame} or a \code{list}.}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element.}

\item{scale}{Whether to scale the data by the scaling factor present in the
file. Defaults to \code{TRUE}.}
}
\value{
A nested list of elements from the specified \code{file}, where the
top levels are chromatograms, peak tables, and/or mass spectra according to
the value of \code{what}. Chromatograms are returned in the format specified
by \code{format_out} (retention time x wavelength).
}
\description{
Reads 'Shimadzu' ascii files into R. These files can be exported from
'Shimadzu LabSolutions' by right clicking on samples in the sample list and
selecting
}
\author{
Ethan Bass
}
