% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_shimadzu}
\alias{read_shimadzu}
\title{Shimadzu ascii reader}
\usage{
read_shimadzu(
  file,
  format_in,
  read_metadata = TRUE,
  format_out = c("matrix", "data.frame"),
  what = c("chromatogram", "peak_table", "both")
)
}
\arguments{
\item{file}{path to file}

\item{format_in}{Format of files. \code{fid} or \code{dad}.}

\item{read_metadata}{Whether to read metadata from file.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{what}{Whether to extract \code{chromatogram}, \code{peak_table} or
\code{both}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Shimadzu ascii reader
}
\author{
Ethan Bass
}
