% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_lcd.R
\name{read_shimadzu_lcd}
\alias{read_shimadzu_lcd}
\title{Read 3D PDA data stream from 'Shimadzu' LCD file.}
\usage{
read_shimadzu_lcd(
  path,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  read_metadata = TRUE
)
}
\arguments{
\item{path}{Path to LCD file.}

\item{format_out}{Matrix or data.frame.}

\item{data_format}{Either \code{wide} (default) or \code{long}.}

\item{read_metadata}{Logical. Whether to attach metadata.}
}
\description{
Relies on the \href{https://pypi.org/project/olefile/}{olefile} package in Python
to unpack files.
}
\note{
This parser is experimental and may still need some work.
}
\author{
Ethan Bass
}
