% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_lcd.R
\name{read_shimadzu_lcd}
\alias{read_shimadzu_lcd}
\title{Read 'Shimadzu' LCD}
\usage{
read_shimadzu_lcd(path)
}
\arguments{
\item{path}{Path to LCD file.}
}
\description{
Read 'Shimadzu' LCD
}
\author{
Ethan Bass
}
