% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_qgd.R
\name{read_shimadzu_qgd}
\alias{read_shimadzu_qgd}
\title{Read 'Shimadzu' QGD}
\usage{
read_shimadzu_qgd(
  path,
  what = c("tic", "ms"),
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw")
)
}
\arguments{
\item{path}{Path to 'Shimadzu' QGD file.}

\item{what}{What stream to get: current options are \code{tic} or
\code{ms}. If a stream is not specified, the function will return both
streams.}

\item{format_out}{Matrix or data.frame.}

\item{data_format}{Either \code{wide} (default) or \code{long}.}

\item{read_metadata}{Logical. Whether to attach metadata.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}
}
\value{
A 2D chromatogram from the chromatogram stream in \code{matrix} or
\code{data.frame} format, according to the value of \code{format_out}.
The chromatograms will be returned in \code{wide} or \code{long} format
according to the value of \code{data_format}.
}
\description{
Read 'Shimadzu' QGD
}
\note{
This parser is experimental and may still need some work. It is not
yet able to interpret much metadata from the files.
}
\author{
Ethan Bass
}
