% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shimadzu_lcd.R
\name{read_sz_qtof_tic}
\alias{read_sz_qtof_tic}
\title{Read 'Shimadzu' QTOF TIC stream}
\usage{
read_sz_qtof_tic(
  path,
  format_out = "data.frame",
  data_format = c("wide", "long"),
  read_metadata = TRUE,
  metadata_format = "shimadzu_lcd"
)
}
\arguments{
\item{path}{Path to 'Shimadzu' \code{.lcd} file.}

\item{format_out}{Matrix or data.frame.}

\item{data_format}{Either \code{wide} (default) or \code{long}.}

\item{read_metadata}{Logical. Whether to attach metadata.}
}
\value{
A 2D chromatogram from the SumTIC stream in \code{matrix} or
\code{data.frame} format, according to the value of \code{format_out}.
The chromatograms will be returned in \code{wide} or \code{long} format
according to the value of \code{data_format}.
}
\description{
A parser to read total ion chromatogram data streams from the quadrupole
time-of-flight (\verb{QTFL RawData}) container of 'Shimadzu' \code{.lcd} files. Triple
quadrupole files store their TIC differently and are read by
\code{read_sz_tlm_tic} instead. LCD files are encoded as 'Microsoft' OLE
documents. The parser relies on the
\href{https://pypi.org/project/olefile/}{olefile} package in Python to unpack the
files. The TIC data is encoded in a stream called \verb{Centroid SumTIC}.
The TIC data stream contains a segment for each retention time, beginning
with a 8-byte header. After the header, the file consists of a series of
4-byte little-endian integers in blocks of 3 (16-bytes per block), followed by
a 4-byte spacer (\code{00000000}) The first integer is the retention time
in milliseconds, the second integer is the scan number, and the third integer
is the intensity. Retention times are converted to minutes, as elsewhere in
the package.
}
\author{
Ethan Bass
}
\keyword{internal}
