% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_thermoraw.R
\name{read_thermoraw}
\alias{read_thermoraw}
\title{Read Thermoraw files into R using ThermoRawFileParser}
\usage{
read_thermoraw(path_in, path_out)
}
\arguments{
\item{path_in}{path to file}

\item{path_out}{directory to export \code{mzML} files.}
}
\value{
A chromatograms in \code{matrix} format.
}
\description{
Parser for thermoraw files
}
\examples{
\dontrun{
read_thermoraw(path)
}
}
\author{
Ethan Bass
}
