% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peaklist.R
\name{read_varian_peaklist}
\alias{read_varian_peaklist}
\title{Read varian peaklist.
Read peak list from 'Varian MS Workstation'.}
\usage{
read_varian_peaklist(file)
}
\arguments{
\item{file}{Path to Varian peak list file.}
}
\description{
Read varian peaklist.
Read peak list from 'Varian MS Workstation'.
}
\author{
Ethan Bass
}
