% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varian_peaklist.R
\name{read_varian_peaklist}
\alias{read_varian_peaklist}
\title{Read 'Varian' peaklist.
Read peak list(s) from 'Varian MS Workstation'.}
\usage{
read_varian_peaklist(file)
}
\arguments{
\item{file}{Path to Varian peak list file.}
}
\description{
Read 'Varian' peaklist.
Read peak list(s) from 'Varian MS Workstation'.
}
\author{
Ethan Bass
}
