% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varian_sms.R
\name{read_varian_sms}
\alias{read_varian_sms}
\title{Read 'Varian' SMS}
\usage{
read_varian_sms(
  path,
  what = c("chrom", "MS1"),
  format_out = c("matrix", "data.frame"),
  data_format = c("long", "wide"),
  read_metadata = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{path}{Path to \code{.SMS} files.}

\item{what}{Whether to extract chromatograms (\code{chroms}) and/or
\code{MS1} data. Accepts multiple arguments.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}

\item{data_format}{Whether to return data in \code{wide} or \code{long} format.}

\item{read_metadata}{Whether to read metadata from file. (This is just a
placeholder for now as there is not yet support for parsing metadata).}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element.}
}
\value{
A nested list of elements from the specified \code{file}, where the
top levels are chromatograms and/or mass spectra according to
the value of \code{what}. Chromatograms are returned in the format specified
by \code{format_out} (retention time x wavelength).
}
\description{
Parser for reading 'Varian Workstation' SMS file.
}
\details{
Varian SMS files begin with a "DIRECTORY" with offsets for each section. The
first section (in all the files I've been able to inspect) is "MSData"
generally beginning at byte 3238. This MSdata section is in turn divided into
two sections. The first section (after a short header) contains chromatogram
data. Some of the information found in this section includes scan numbers,
retention times, (as 64-bit
floats), the total ion chromatogram (TIC), the base peak chromatogram (BPC),
as well as some other unidentified information. The scan numbers and
intensities for the TIC and BPC are stored at 4-byte little-endian integers.
Following this section, there is a series of null bytes, followed by a series
of segments containing the mass spectra.

The encoding scheme for the mass spectra is somewhat more complicated. Each
scan is represented by a series of values of variable length separated from
the next scan by two null bytes. Within these segments, values are paired.
The first value in each pair represents the delta-encoded mass-to-charge ratio,
while the second value represents the intensity of the signal. All values in
this section are offset encoded using a base of value of 4096 (16^3). Thus,
integers beginning with digits 0-3 are simple 2-byte integers. However, values
with a prefix digit >= 4 are 4-byte integers with offset encoding, where the
offset is (n-4) * 4096. So values with a prefix digit of 4 have an offset of 0,
values with a prefix digit of 5 have an offset of 8192 (4096 * 2), and so on.
}
\note{
There is not yet support for the extraction of metadata from this file
format.
}
\author{
Ethan Bass
}
