% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_waters_arw}
\alias{read_waters_arw}
\title{Waters ascii (.arw) reader}
\usage{
read_waters_arw(
  file,
  read_metadata = TRUE,
  format_out = c("matrix", "data.frame")
)
}
\arguments{
\item{file}{path to file}

\item{read_metadata}{Whether to read metadata from file.}

\item{format_out}{R format. Either \code{matrix} or \code{data.frame}.}
}
\value{
A chromatogram in the format specified by \code{format_out}
(retention time x wavelength).
}
\description{
Reads 'Waters ARW' files.
}
\details{
For help exporting files from Empower, you can consult the official
documentation: \href{https://support.waters.com/KB_Inf/Empower_Breeze/WKB77571_How_to_export_3D_raw_data_from_Empower_to_a_Microsoft_Excel_spreadsheet}{How_to_export_3D_raw_data_from_Empower}.
}
\author{
Ethan Bass
}
