% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-roxygen.R
\name{shared_params}
\alias{shared_params}
\title{Shared params 2D chromatogram}
\arguments{
\item{format_out}{Class of output. Either \code{matrix}, \code{data.frame}, or
\code{data.table}.}

\item{data_format}{Whether to return data in \code{wide} (default) or \code{long} format.}

\item{metadata_format}{Format to output metadata. Either \code{chromconverter}
or \code{raw}.}

\item{progress_bar}{Logical. Whether to show progress bar. Defaults to \code{TRUE}
if \code{pbapply} is installed.}

\item{cl}{Argument to \link[pbapply:pbapply]{pbapply} specifying the number
of clusters to use or a cluster object created by
\link[parallel:makeCluster]{makeCluster}. Defaults to \code{1}.}

\item{read_metadata}{Logical. Whether to attach metadata. Defaults to \code{TRUE}.}

\item{collapse}{Logical. Whether to collapse lists that only contain a single
element. Defaults to \code{TRUE}.}

\item{scale}{Whether to scale the data by the scaling factor present in the
file. Defaults to \code{TRUE}.}
}
\value{
A chromatogram in the format specified by the \code{format_out} and
\code{data_format} arguments.
}
\description{
Shared params 2D chromatogram
}
\keyword{internal}
