% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chroms.R
\name{sp_converter}
\alias{sp_converter}
\title{converter for Agilent MassHunter UV files}
\usage{
sp_converter(file)
}
\arguments{
\item{file}{path to file}
}
\value{
A data.frame object (retention time x wavelength).
}
\description{
MassHunter converter
}
\details{
Converts a single chromatogram from MassHunter \code{.sp} format to R \code{data.frame}.
}
