% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chroms.R
\name{uv_converter}
\alias{uv_converter}
\title{converter for Agilent Chemstation UV files}
\usage{
uv_converter(file)
}
\arguments{
\item{file}{path to file}
}
\value{
A data.frame object (retention time x trace).
}
\description{
ChemStation converter
}
\details{
Converts a single chromatogram from Chemstation \code{.uv} format to R \code{data.frame}.
}
