% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chroms.R
\name{uv_converter}
\alias{uv_converter}
\title{Converter for Agilent ChemStation UV files}
\usage{
uv_converter(file)
}
\arguments{
\item{file}{path to file}
}
\value{
A data.frame object (retention time x trace).
}
\description{
Converts a single chromatogram from ChemStation \code{.uv} format to R \code{data.frame}.
}
\details{
Uses the \href{https://github.com/bovee/aston}{Aston} file parser.
}
