% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_chroms.R
\name{write_chroms}
\alias{write_chroms}
\title{Write chromatograms}
\usage{
write_chroms(
  chrom_list,
  path_out,
  export_format = c("mzml", "cdf", "csv", "arw"),
  what = "",
  force = FALSE,
  show_progress = TRUE,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{chrom_list}{A list of chromatograms.}

\item{path_out}{Path to directory for writing files.}

\item{export_format}{Format to export files: either \code{"mzml"}, \code{"cdf"},
\code{"csv"}, \code{"arw"}.}

\item{what}{What to write. Used by the \code{cdf} and \code{mzml} exporters only.
For \code{export_format = "cdf"}, either \code{"MS1"} (written by \code{write_andi_ms}) or
\code{"chrom"} (written by \code{write_andi_chrom}). For \code{export_format = "mzml"},
any of \code{"MS1"}, \code{"MS2"}, \code{"TIC"}, \code{"BPC"} and/or \code{"DAD"}. If it is not
specified, the streams to write are inferred from the supplied data.}

\item{force}{Logical. Whether to overwrite existing files. Defaults to \code{TRUE}.}

\item{show_progress}{Logical. Whether to show progress bar. Defaults to \code{TRUE}.}

\item{verbose}{Logical. Whether to print verbose output.}

\item{...}{Additional arguments to write function.}
}
\value{
No return value. The function is called for its side effects.
}
\description{
Writes chromatograms to disk in the format specified by \code{export_format}:
either \code{mzml}, \code{cdf}, \code{csv}, or \code{arw}.
}
\section{Side effects}{

Exports a chromatogram in the file format specified by \code{export_format} in the
directory specified by \code{path_out}.
}

\examples{
path <- system.file("extdata/ladder.txt", package = "chromConverter")
chroms <- read_chroms(path, format_in = "shimadzu_ascii",
                      find_files = FALSE, progress_bar = FALSE)
write_chroms(chroms, path_out = tempdir(), export_format = "csv",
             show_progress = FALSE)
}
\seealso{
Other write functions:
\code{\link[=write_andi_chrom]{write_andi_chrom()}},
\code{\link[=write_andi_ms]{write_andi_ms()}},
\code{\link[=write_mzml]{write_mzml()}}
}
\author{
Ethan Bass
}
\concept{write functions}
